requirements.txt
requests
beautifulsoup4
langchain
langchain-openai
python-dotenv
tweepy
openai
faiss-cpu
pytz
langgraph>=0.2.0
pydantic[email]
flask
werkzeug